bookings, nodemailer, qrcode, & cloudinary done

+++ controllers/bookingController.js
@@ -2,13 +2,15 @@ const { Booking, Movie } = require("../models");
 const transporter = require("../utils/mail");
 const createMailOptions = require("../utils/createMailOptions");
 const nodemailer = require("nodemailer");
+const axios = require("axios");
+const generateQR = require("../utils/generateQR");
 require("dotenv").config();
+const cloudinary = require("cloudinary");
 
 class BookingController {
   static async addBooking(req, res, next) {
     try {
       //check if movie exist
-      console.log(req.params.mid);
       const isMovie = await Movie.findByPk(+req.params.mid);
       if (!isMovie) {
         throw { message: "Movie not found" };
@@ -25,9 +27,16 @@ class BookingController {
       if (!created) {
         throw { message: "Already Booked" };
       }
+      //qrcode
+      const qrcode = await generateQR(`${isMovie.imdbUrl}`);
+      //cloudinary
+      cloudinary.config({
+        cloud_name: process.env.CLOUDINARY_USER_NAME,
+        api_key: process.env.CLOUDINARY_API_KEY,
+        api_secret: process.env.CLOUDINARY_API_SECRET,
+      });
+      const uploadResponse = await cloudinary.uploader.upload(qrcode, {});
       //nodemailer
-      //   console.log(isMovie.title);
-      //   console.log(req.user);
       function sendMessage() {
         try {
           // mail options
@@ -39,6 +48,7 @@ class BookingController {
                   <p>Actors: ${isMovie.actors}</p>
                   <img src=${isMovie.poster}>
                   <p>Plot: ${isMovie.plot}</p>
+                  <img src="${uploadResponse.url}">
                   <br/>
                 `;
           const mailOptions = {
@@ -51,6 +61,7 @@ class BookingController {
           transporter.sendMail(mailOptions, function (err, info) {
             if (err) {
               console.log("Error sending message: " + err);
+              throw { message: `Nodemailer Failed` };
             } else {
               // no errors, it worked
               console.log("Message sent succesfully.");
@@ -60,7 +71,6 @@ class BookingController {
           console.log("Other error sending message: " + error);
         }
       }
-      // thats the key part, without all these it didn't work for me
       let transporter = nodemailer.createTransport({
         host: "smtp.gmail.com",
         port: 465,
@@ -68,13 +78,12 @@ class BookingController {
         service: "gmail",
         auth: {
           type: "OAUTH2",
-          user: process.env.GMAIL_USERNAME, //set these in your .env file
+          user: process.env.GMAIL_USERNAME,
           clientId: process.env.OAUTH_CLIENT_ID,
           clientSecret: process.env.OAUTH_CLIENT_SECRET,
           refreshToken: process.env.OAUTH_REFRESH_TOKEN,
         },
       });
-
       // invoke sending function
       sendMessage();
       //response
@@ -107,10 +116,67 @@ class BookingController {
 
   static async deleteBooking(req, res, next) {
     try {
+      //check if movie exist
+      const isMovie = await Movie.findByPk(+req.params.mid);
+      if (!isMovie) {
+        throw { message: "Movie not found" };
+      }
       const cancel = await Booking.destroy({
         where: { UserId: req.user.id, MovieId: +req.params.mid },
         truncate: false,
       });
+      //nodemailer
+      //   console.log(isMovie.title);
+      //   console.log(req.user);
+      function sendMessage() {
+        try {
+          // mail options
+          const htmlContent = `
+                  <h1><strong>Booking Cancelled -- CINTIA</strong></h1>
+                  <p>Hi, You have cancelled this booking:</p>
+                  <h2>${isMovie.title}</h2>
+                  <p>Genre: ${isMovie.genre}</p>
+                  <p>Actors: ${isMovie.actors}</p>
+                  <img src=${isMovie.poster}>
+                  <p>Plot: ${isMovie.plot}</p>
+                  <br/>
+                `;
+          const mailOptions = {
+            from: "CINTIA Cinema Ticketing Apps",
+            to: `${req.user.email}`,
+            subject: "Your Booking Info",
+            html: htmlContent,
+          };
+          // here we actually send it
+          transporter.sendMail(mailOptions, function (err, info) {
+            if (err) {
+              console.log("Error sending message: " + err);
+              throw { message: `Nodemailer Failed` };
+            } else {
+              // no errors, it worked
+              console.log("Message sent succesfully.");
+            }
+          });
+        } catch (error) {
+          console.log("Other error sending message: " + error);
+        }
+      }
+      // thats the key part, without all these it didn't work for me
+      let transporter = nodemailer.createTransport({
+        host: "smtp.gmail.com",
+        port: 465,
+        secure: true,
+        service: "gmail",
+        auth: {
+          type: "OAUTH2",
+          user: process.env.GMAIL_USERNAME, //set these in your .env file
+          clientId: process.env.OAUTH_CLIENT_ID,
+          clientSecret: process.env.OAUTH_CLIENT_SECRET,
+          refreshToken: process.env.OAUTH_REFRESH_TOKEN,
+        },
+      });
+      // invoke sending function
+      sendMessage();
       res.status(200).json({ message: `Booking Successfully Cancelled` });
     } catch (err) {
       next(err);
